Added all images locally
I had to do this as there was a CORs issue when trying to send http requests from JS to the CW's image server. I hope it works this time as I can use github's server ;)

PS: i made a PY script that automatically downloaded the images and stored them into the 'images' folder. the PY script will be listed.
Co-Authored-By: Zach <196121962+zmushtare@users.noreply.github.com>
Co-Authored-By: Imgbot <help@imgbot.net>